# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Aeza Server API
  version: 1.0.0
  description: REST, WebSocket and SSE API for frontend and agents.
servers:
  - url: http://158.160.205.79:8087
    description: Current deployed server
  - url: /
    description: Relative current server
paths:
...
  /api/v1/logs/search:
    get:
      summary: Search logs
      description: Returns logs with total count and aggregations for frontend filters, including agent-side metadata and agent group filtering.
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
          description: Full-text query over message, host, service, sourcePath, messageTemplate and fingerprint.
        - in: query
          name: host
          required: false
          schema:
            type: string
        - in: query
          name: service
          required: false
          schema:
            type: string
        - in: query
          name: level
          required: false
          schema:
            type: string
        - in: query
          name: agentId
          required: false
          schema:
            type: string
        - in: query
          name: groupId
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: fingerprint
          required: false
          schema:
            type: string
        - in: query
          name: aggregated
          required: false
          schema:
            type: boolean
        - in: query
          name: sampled
          required: false
          schema:
            type: boolean
        - in: query
          name: burstDetected
          required: false
          schema:
            type: boolean
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: groupId
          required: false
          schema:
            type: integer
            format: int64
        - in: query
          name: page
          required: false
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Search response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
              example:
                items:
                  - timestamp: "2026-03-21T11:33:50Z"
                    level: "ERROR"
                    message: "database failed 2"
                    host: "agent-vm-01"
                    service: "nginx"
                    sourceType: "file"
                    sourcePath: "/var/log/nginx/error.log"
                    tags: ["prod", "agent"]
                    metadata:
                      fingerprint: "d5c98f0a8be1d4ac"
                      messageTemplate: "database failed <num>"
                      occurrences: 12
                      aggregated: true
                      burstDetected: true
                    agentId: "agent-dev-01"
                total: 1
                aggregations:
                  hosts:
                    agent-vm-01: 1
                  services:
                    nginx: 1
                  levels:
                    ERROR: 1
                  fingerprints:
                    d5c98f0a8be1d4ac: 1
...
  /api/v1/agents/heartbeat:
    post:
      summary: Update agent heartbeat
      description: Agent endpoint for updating current status, buffered count and watched sources.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatRequest"
            example:
              agentId: "agent-dev-01"
              agentName: "Main dev agent"
              hostName: "agent-vm-01"
              hostIp: "194.113.106.38"
              status: "online"
              bufferedCount: 0
              lastSequence: 1774104589716
              watchedSources:
                - "/var/log/nginx/error.log"
                - "/var/log/syslog"
      responses:
        "200":
          description: Agent status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
              example:
                agentId: "agent-dev-01"
                name: "Main dev agent"
                host: "agent-vm-01"
                hostIp: "194.113.106.38"
                status: "online"
                lastSeen: "2026-03-21T11:24:32.636001841Z"
                bufferedCount: 0
                groupId: 1
                groupName: "production"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                timestamp: "2026-03-21T12:00:00Z"
                status: 400
                error: "Bad Request"
                message: "validation failed"
                path: "/api/v1/agents/heartbeat"
  /api/v1/ingest/batch:
    post:
      summary: Ingest batch of logs
      description: Agent endpoint for sending a batch of logs.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestBatchRequest"
            example:
              agentId: "agent-dev-01"
              sequence: 1774104589716
              events:
                - timestamp: "2026-03-21T11:40:00Z"
                  level: "ERROR"
                  message: "database timeout"
                  host: "agent-vm-01"
                  service: "nginx"
                  sourceType: "file"
                  sourcePath: "/var/log/nginx/error.log"
                  tags: ["prod", "agent"]
                  metadata:
                    env: "prod"
                    component: "collector"
      responses:
        "200":
          description: Ingest accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IngestAckResponse"
              example:
                status: "ok"
                accepted: 1
                requestId: "27aa717b-688f-458c-954f-f8fa8d8cbd01"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
              example:
                timestamp: "2026-03-21T12:00:00Z"
                status: 400
                error: "Bad Request"
                message: "validation failed"
                path: "/api/v1/ingest/batch"
...
components:
  schemas:
...
    LogSearchItem:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
        message:
          type: string
        host:
          type: string
        service:
          type: string
        sourceType:
          type: string
          nullable: true
        sourcePath:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        agentId:
          type: string
    SearchResponse:
      type: object
      required:
        - items
        - total
        - aggregations
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LogSearchItem"
        total:
          type: integer
          format: int64
        aggregations:
          type: object
          additionalProperties: true
...
    AgentStatus:
      type: object
      properties:
        agentId:
          type: string
        name:
          type: string
        host:
          type: string
        hostIp:
          type: string
          nullable: true
        status:
          type: string
        lastSeen:
          type: string
          format: date-time
        bufferedCount:
          type: integer
        groupId:
          type: integer
          format: int64
          nullable: true
        groupName:
          type: string
          nullable: true
...
    HeartbeatRequest:
      type: object
      required:
        - agentId
        - hostName
        - status
        - bufferedCount
      properties:
        agentId:
          type: string
        agentName:
          type: string
        hostName:
          type: string
        hostIp:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - online
            - degraded
        bufferedCount:
          type: integer
          minimum: 0
        lastSequence:
          type: integer
          format: int64
        watchedSources:
          type: array
          items:
            type: string
    LogEvent:
      type: object
      required:
        - timestamp
        - level
        - message
        - host
        - service
      properties:
        timestamp:
          type: string
          format: date-time
        level:
          type: string
        message:
          type: string
        host:
          type: string
        service:
          type: string
        sourceType:
          type: string
          nullable: true
        sourcePath:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        metadata:
          type: object
          additionalProperties: true
        agentId:
          type: string
          nullable: true
    IngestBatchRequest:
      type: object
      required:
        - agentId
        - sequence
        - events
      properties:
        agentId:
          type: string
        sequence:
          type: integer
          format: int64
        events:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/LogEvent"
    IngestAckResponse:
      type: object
      properties:
        status:
          type: string
        accepted:
          type: integer
        requestId:
          type: string
...
    ApiErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string